refactor types, function, and argument naming
