--- src/emoji.rs
@@ -1,11 +1,0 @@
-//! Emoji contants used by `prjfmt`.
-
-#![allow(missing_docs)]
-
-use console::Emoji;
-
-pub static FOLDER: Emoji = Emoji("📂", "");
-pub static WARN: Emoji = Emoji("⚠️", ":-)");
-pub static ERROR: Emoji = Emoji("⛔", "");
-pub static INFO: Emoji = Emoji("ℹ️", "");
-pub static DEBUG: Emoji = Emoji("🐛", "");
