+++ src/formatters/manifest.rs
@@ -1,23 +1,21 @@
-use crate::{emoji, CLOG};
+use super::RootManifest;
+use crate::{customlog, CmdContext, CLOG};
 
-use crate::engine::CmdContext;
 use anyhow::{anyhow, Result};
 use hex;
-use serde::{Deserialize, Serialize};
 use sha1::{Digest, Sha1};
 use std::collections::BTreeMap;
 use std::fs::{read_to_string, File};
 use std::io::Write;
 use std::path::PathBuf;
-use std::str;
 
 /// Create <hex(hash(path-to-prjfmt))>.toml and put it in $XDG_CACHE_DIR/prjfmt/eval-cache/
-pub fn create_prjfmt_manifest(
+pub fn create_manifest(
     prjfmt_toml: PathBuf,
     cache_dir: PathBuf,
     cmd_ctx: Vec<CmdContext>,
 ) -> Result<()> {
-    let hash_toml = create_prjfmt_hash(&prjfmt_toml)?;
+    let hash_toml = create_hash(&prjfmt_toml)?;
 
     let mut f = File::create(cache_dir.as_path().join(hash_toml))?;
     let map_manifest: BTreeMap<String, CmdContext> = cmd_ctx
@@ -38,8 +36,8 @@ pub fn create_prjfmt_manifest(
     f.write_all(
         format!(
             "# {} DO NOT HAND EDIT THIS FILE {}\n\n{}",
-            emoji::WARN,
-            emoji::WARN,
+            customlog::WARN,
+            customlog::WARN,
             toml::to_string_pretty(&manifest_toml)?
         )
         .as_bytes(),
@@ -48,12 +46,12 @@ pub fn create_prjfmt_manifest(
 }
 
 /// Read the <hex(hash(path-to-prjfmt))>.toml and return list of config's cache evaluation
-pub fn read_prjfmt_manifest(prjfmt_toml: &PathBuf, path: &PathBuf) -> Result<RootManifest> {
-    let hash_toml = create_prjfmt_hash(&prjfmt_toml)?;
-    let manifest_toml = path.as_path().join(&hash_toml);
+pub fn read_manifest(prjfmt_toml: &PathBuf, cache_dir: &PathBuf) -> Result<RootManifest> {
+    let hash_toml = create_hash(&prjfmt_toml)?;
+    let manifest_toml = cache_dir.as_path().join(&hash_toml);
 
     if manifest_toml.as_path().exists() {
-        CLOG.debug(&format!("Found {} in: {}", hash_toml, path.display()));
+        CLOG.debug(&format!("Found {} in: {}", hash_toml, cache_dir.display()));
         let open_file = match read_to_string(manifest_toml.as_path()) {
             Ok(file) => file,
             Err(err) => {
@@ -75,22 +73,20 @@ pub fn read_prjfmt_manifest(prjfmt_toml: &PathBuf, path: &PathBuf) -> Result<Roo
     }
 }
 
-fn create_prjfmt_hash(prjfmt_toml: &PathBuf) -> Result<String> {
+fn create_hash(prjfmt_toml: &PathBuf) -> Result<String> {
     let prjfmt_str = match prjfmt_toml.to_str() {
         Some(str) => str.as_bytes(),
-        None => return Err(anyhow!("{}cannot convert to string slice", emoji::ERROR)),
+        None => {
+            return Err(anyhow!(
+                "{}cannot convert to string slice",
+                customlog::ERROR
+            ))
+        }
     };
     let prjfmt_hash = Sha1::digest(prjfmt_str);
     let result = hex::encode(prjfmt_hash);
     let manifest_toml = format!("{}.toml", result);
     Ok(manifest_toml)
-}
-
-#[derive(Debug, Deserialize, Serialize)]
-/// RootManifest
-pub struct RootManifest {
-    /// Map of manifests config based on its formatter
-    pub manifest: BTreeMap<String, CmdContext>,
 }
 
 #[cfg(test)]
@@ -99,10 +95,10 @@ mod tests {
 
     /// Every same path produce same hash
     #[test]
-    fn test_create_prjfmt_hash() -> Result<()> {
+    fn test_create_hash() -> Result<()> {
         let file_path = PathBuf::from(r"examples/monorepo/prjfmt.toml");
         let prjfmt_hash = "02e97bc0a67b5d61f3152c184690216085ef0c03.toml";
-        assert_eq!(create_prjfmt_hash(&file_path)?, prjfmt_hash);
+        assert_eq!(create_hash(&file_path)?, prjfmt_hash);
         Ok(())
     }
 }
